+++ tests/StackExchange.Redis.Tests/ExceptionFactoryTests.cs
@@ -124,6 +124,7 @@ public class ExceptionFactoryTests : TestBase
             Assert.Contains("sync-ops: ", ex.Message);
             Assert.Contains("async-ops: ", ex.Message);
             Assert.Contains("conn-sec: n/a", ex.Message);
+            Assert.Contains("aoc: 1", ex.Message);
 #if NETCOREAPP
                 Assert.Contains("POOL: ", ex.Message);
 #endif